MGDOBR-865: Run InstanceTest (partialy) with Dev Stage
