+++ cypress/cypress.e2e.json
@@ -14,6 +14,7 @@
   },
   "viewportWidth": 1920,
   "viewportHeight": 1080,
+  "defaultCommandTimeout": 10000,
   "env": {
     "TEST_TYPE": "dev"
   }
